requirements.txt
python-dotenv
groq
requests
